# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/register/follower": {
      "post": {
        "operationId": "AuthController_registerFollower",
        "summary": "Register a follower (contract §4.1)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterFollowerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/register/creator": {
      "post": {
        "operationId": "AuthController_registerCreator",
        "summary": "Register a creator; opens a pending verification application (contract §4.1)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCreatorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/otp/request": {
      "post": {
        "operationId": "AuthController_otpRequest",
        "summary": "Request an OTP for a phone number (contract §4.1)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_otpVerify",
        "summary": "Verify an OTP and sign in (contract §4.1)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Email + password login (contract §4.1)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password/forgot": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "summary": "Start a password reset; emails a token (contract §4.1)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password/reset": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "summary": "Complete a password reset with the emailed token (contract §4.1)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "ok": true
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password/change": {
      "post": {
        "operationId": "AuthController_changePassword",
        "summary": "Change password for the logged-in user (contract §4.1)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "ok": true
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/auth/social/apple": {
      "post": {
        "operationId": "AuthController_apple",
        "summary": "Sign in with Apple (contract §4.1)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleSocialDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/social/google": {
      "post": {
        "operationId": "AuthController_google",
        "summary": "Sign in with Google (contract §4.1)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleSocialDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Log out (contract §4.1). Stateless JWT — client discards the token.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/alerts": {
      "get": {
        "operationId": "AlertsController_list",
        "summary": "My price alerts (contract §alerts)",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "active",
                "triggered",
                "cancelled"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PriceAlertDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "alerts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "AlertsController_create",
        "summary": "Create a price alert (contract §alerts)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePriceAlertDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceAlertDto"
                }
              }
            }
          }
        },
        "tags": [
          "alerts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/alerts/{id}": {
      "delete": {
        "operationId": "AlertsController_cancel",
        "summary": "Cancel a price alert (contract §alerts)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "alerts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Millimore API",
    "description": "Backend API for the Millimore platform (mobile app + admin dashboard). Single source of truth: docs/BACKEND_CONTRACT.md",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.millimore.app/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterFollowerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Priya Sharma"
          },
          "phone": {
            "type": "string",
            "example": "+91 90000 00000"
          },
          "residenceIso": {
            "type": "string",
            "example": "IN"
          },
          "experience": {
            "type": "string",
            "example": "beginner"
          },
          "interests": {
            "example": [
              "Forex",
              "Crypto"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "photoUrl": {
            "type": "string"
          },
          "referralCode": {
            "type": "string",
            "example": "A1B2C3D",
            "description": "Referral code the user signed up with."
          }
        },
        "required": [
          "name",
          "phone",
          "residenceIso"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "u_123"
          },
          "name": {
            "type": "string",
            "example": "Marcus Sterling"
          },
          "username": {
            "type": "string",
            "example": "marcussterling"
          },
          "email": {
            "type": "object",
            "example": "marcus@x.com",
            "nullable": true
          },
          "phone": {
            "type": "object",
            "example": "+91 90000 00000",
            "nullable": true
          },
          "photoUrl": {
            "type": "object",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "follower",
              "creator",
              "admin"
            ]
          },
          "creatorStatus": {
            "type": "string",
            "enum": [
              "none",
              "pending",
              "approved",
              "rejected",
              "suspended"
            ]
          },
          "residenceIso": {
            "type": "object",
            "example": "IN",
            "nullable": true
          },
          "residenceCountry": {
            "type": "object",
            "example": "India",
            "nullable": true
          },
          "market": {
            "type": "object",
            "example": "Forex",
            "nullable": true
          },
          "platform": {
            "type": "object",
            "example": "MetaTrader 5",
            "nullable": true
          },
          "addressLine": {
            "type": "object",
            "nullable": true
          },
          "city": {
            "type": "object",
            "nullable": true
          },
          "postalCode": {
            "type": "object",
            "nullable": true
          },
          "leverage": {
            "type": "number",
            "example": 100,
            "description": "Default leverage."
          },
          "kycStatus": {
            "type": "string",
            "enum": [
              "none",
              "pending",
              "verified",
              "rejected"
            ],
            "description": "KYC status."
          },
          "twofaEnabled": {
            "type": "boolean",
            "example": false,
            "description": "Whether 2FA is enabled."
          },
          "createdAt": {
            "type": "string",
            "example": "2026-01-01T00:00:00Z"
          }
        },
        "required": [
          "id",
          "name",
          "username",
          "role",
          "creatorStatus",
          "leverage",
          "kycStatus",
          "twofaEnabled",
          "createdAt"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "required": [
          "token",
          "user"
        ]
      },
      "VerificationDto": {
        "type": "object",
        "properties": {
          "platform": {
            "type": "string",
            "example": "MetaTrader 5"
          },
          "server": {
            "type": "string",
            "example": "XM-Live3"
          },
          "account": {
            "type": "string",
            "example": "50231487"
          },
          "investorPassword": {
            "type": "string",
            "example": "investor-pw",
            "writeOnly": true
          },
          "statementUrl": {
            "type": "string",
            "example": "https://storage/statement.pdf"
          }
        },
        "required": [
          "platform"
        ]
      },
      "RegisterCreatorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Marcus Sterling"
          },
          "phone": {
            "type": "string",
            "example": "+91 90000 00000"
          },
          "residenceIso": {
            "type": "string",
            "example": "IN"
          },
          "market": {
            "type": "string",
            "example": "Forex"
          },
          "platform": {
            "type": "string",
            "example": "MetaTrader 5"
          },
          "verification": {
            "$ref": "#/components/schemas/VerificationDto"
          },
          "referralCode": {
            "type": "string",
            "example": "A1B2C3D",
            "description": "Referral code the user signed up with."
          }
        },
        "required": [
          "name",
          "phone",
          "residenceIso",
          "market",
          "platform",
          "verification"
        ]
      },
      "OtpRequestDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+91 90000 00000"
          }
        },
        "required": [
          "phone"
        ]
      },
      "OtpRequestResponseDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "example": "otp_9f8a..."
          },
          "devCode": {
            "type": "string",
            "description": "Present only in dev/console OTP mode to ease app integration.",
            "example": "123456"
          }
        },
        "required": [
          "requestId"
        ]
      },
      "OtpVerifyDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "example": "otp_9f8a..."
          },
          "code": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "requestId",
          "code"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "trader@millimore.app"
          },
          "password": {
            "type": "string",
            "example": "password",
            "writeOnly": true
          },
          "twofaCode": {
            "type": "string",
            "example": "287082",
            "description": "2FA code (or backup code) when 2FA is enabled."
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "trader@millimore.app"
          }
        },
        "required": [
          "email"
        ]
      },
      "ForgotPasswordResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "devToken": {
            "type": "string",
            "nullable": true,
            "description": "Dev-only: the reset token, surfaced when MAIL_PROVIDER=console (never in production)."
          }
        },
        "required": [
          "ok"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "The token from the reset email."
          },
          "newPassword": {
            "type": "string",
            "example": "new-strong-password",
            "writeOnly": true,
            "minLength": 8
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "writeOnly": true
          },
          "newPassword": {
            "type": "string",
            "example": "new-strong-password",
            "writeOnly": true,
            "minLength": 8
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "AppleSocialDto": {
        "type": "object",
        "properties": {
          "identityToken": {
            "type": "string",
            "description": "Apple identity token (JWT)."
          }
        },
        "required": [
          "identityToken"
        ]
      },
      "GoogleSocialDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "description": "Google ID token (JWT)."
          }
        },
        "required": [
          "idToken"
        ]
      },
...
      "PriceAlertDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "pa_1"
          },
          "symbol": {
            "type": "string",
            "example": "XAU/USD"
          },
          "direction": {
            "type": "string",
            "enum": [
              "above",
              "below"
            ]
          },
          "targetPrice": {
            "type": "number",
            "example": 2450
          },
          "note": {
            "type": "object",
            "nullable": true,
            "example": "Take profit zone"
          },
          "notifyEmail": {
            "type": "boolean",
            "example": false,
            "description": "Also email when it triggers."
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "triggered",
              "cancelled"
            ]
          },
          "triggeredAt": {
            "type": "object",
            "nullable": true
          },
          "triggeredPrice": {
            "type": "object",
            "nullable": true,
            "example": 2450.2
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "symbol",
          "direction",
          "targetPrice",
          "notifyEmail",
          "status",
          "createdAt"
        ]
      },
      "CreatePriceAlertDto": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string",
            "example": "XAU/USD",
            "description": "A tradable symbol (see GET /symbols)."
          },
          "direction": {
            "type": "string",
            "enum": [
              "above",
              "below"
            ],
            "description": "Fire when price crosses the target upward (above) or downward (below)."
          },
          "targetPrice": {
            "type": "number",
            "example": 2450,
            "description": "Target price; must be positive."
          },
          "note": {
            "type": "string",
            "example": "Take profit zone",
            "maxLength": 140
          },
          "notifyEmail": {
            "type": "boolean",
            "example": false,
            "default": false,
            "description": "Also send an email when it triggers."
          }
        },
        "required": [
          "symbol",
          "direction",
          "targetPrice"
        ]
      },
...
    }
  }
}